+++ cmp/grid/index.js
@@ -9,8 +9,4 @@ export * from './Grid';
 export * from './GridModel';
 export * from './GridStateModel';
 
-export * from './columns/Core';
-export * from './columns/DatesTimes';
-export * from './columns/ExportFormat';
-export * from './columns/Column';
-export * from './columns/ColumnGroup';
+export * from './columns/';