+++ desktop/cmp/grid/columns/index.js
@@ -5,8 +5,4 @@
  * Copyright © 2018 Extremely Heavy Industries Inc.
  */
 
-export * from './impl/ColChooser';
-export * from './impl/ColChooserModel';
-export * from './columns';
-
-
+export * from './Actions';